Support for large numbers Changed to long long in algorithm UI tweaks to support displaying large numbers GCD for prime testing

+++ Other Projects/Primes/Obj-C/Primer/Prime.h
@@ -11,5 +11,5 @@
 @interface Prime : NSObject{
 }
 -(NSMutableArray*)listPrimes: (double)cap;
--(bool)testPrimeness: (int)candidate;
+-(bool)testPrimeness: (long long)candidate;
 @end
